% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{gauss_vec_0_mean_ar1_cov}
\alias{gauss_vec_0_mean_ar1_cov}
\title{This function generate Gaussian Vector with mean 0 and AR1 covariance.
X(t)~N(0,sigma2*rho^{ct}) i.e. the variance component is invariant of t}
\usage{
gauss_vec_0_mean_ar1_cov(t, sigma2, rho, c)
}
\arguments{
\item{t}{grid points for evaluation}

\item{sigma2}{marginal variance}

\item{rho}{correlation coefficient}

\item{c}{correlation coefficient parameter}
}
\value{
Returns The returned vector is X(t_1), X(t_2), ..., X(t_pp)
}
\description{
This function generate Gaussian Vector with mean 0 and AR1 covariance.
X(t)~N(0,sigma2*rho^{ct}) i.e. the variance component is invariant of t
}
