% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{gauss_vec_0_mean_given_cov}
\alias{gauss_vec_0_mean_given_cov}
\title{This function generate Multivariate Normal with mean 0 and given covariance matrix}
\usage{
gauss_vec_0_mean_given_cov(B, m, cov)
}
\arguments{
\item{B}{is the number of iid gaussian vectors we want to return;}

\item{m}{is the size of each gaussian vector;}

\item{cov}{is the given covariance matrix}
}
\value{
Returns a matrix with multivariate normal with mean 0 and given covariance matrix
}
\description{
This function generate Multivariate Normal with mean 0 and given covariance matrix
}
