\name{hw-package}
\alias{hw-package}
\alias{hw}
\docType{package}
\title{
\packageTitle{hw}
}
\description{
\packageDescription{hw}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{hw}
\packageIndices{hw}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{hw}

Maintainer: \packageMaintainer{hw}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
