% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{ker}
\alias{ker}
\title{This function Epanechnikov Kernel function}
\usage{
ker(x, h)
}
\arguments{
\item{x}{vector for evaluation}

\item{h}{bandwidth}
}
\value{
Returns the kernel function values
}
\description{
This function Epanechnikov Kernel function
}
