% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_lin_reg.R
\name{mk_lm}
\alias{mk_lm}
\title{MK_Regression}
\usage{
mk_lm(Y, X, X_names)
}
\arguments{
\item{Y}{outcome variable must be in vector format}

\item{X}{covariate matrix, each column is a different covariate. Note, if only one covariate used it is reccomended to still be a matrix.}

\item{X_names}{the names of the covariates in the model as a vector of string(s)}
}
\value{
a table is returned that specifies the beta estimate, standard error, t statistic, and p value
}
\description{
This function uses ordinary least squares to do linear regression.  It requires an input of Y (outcome) and
X (matrix of covariates).  It will then run a linear regression and poutput the beta coefficiants, standard error,
t statistic, and p value (using a 2 sided test with alpha = 0.05 significance).
}
\examples{
mk_lm(Y = mtcars$mpg, X = cbind(mtcars$wt, mtcars$cyl), X_names = c("weight", "cylinders"))
mk_lm(Y = mtcars$mpg, X = matrix(mtcars$wt, ncol = 1), X_names = c("weight"))

}
