% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{S_half}
\alias{S_half}
\title{This function calculate the sqrt root of a pd matrix}
\usage{
S_half(S)
}
\arguments{
\item{S}{PD matrix}
}
\value{
Returns a aqrt root of the matrix
}
\description{
This function calculate the sqrt root of a pd matrix
}
